ya sale results y recommendations solo falta la gráfica
